+++ src/Service.Tests/Authentication/EasyAuthAuthenticationUnitTests.cs
@@ -2,18 +2,14 @@
 // Licensed under the MIT License.
 
 #nullable enable
-using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Azure.DataApiBuilder.Config.ObjectModel;
 using Azure.DataApiBuilder.Core.AuthenticationHelpers;
 using Azure.DataApiBuilder.Core.Authorization;
-using Azure.DataApiBuilder.Core.Configurations;
 using Azure.DataApiBuilder.Service.Tests.Authentication.Helpers;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.TestHost;
@@ -348,49 +344,6 @@ namespace Azure.DataApiBuilder.Service.Tests.Authentication
         }
 
         /// <summary>
-        /// Tests we ensure that an invalid query in the EasyAuth header does not result in a successful request.
-        /// </summary>
-        [TestMethod]
-        public async Task TestInvalidQueryInHeader()
-        {
-            string header = @"
-            {
-                ""auth_typ"":""aad"",
-                ""claims"":[
-                  {
-                    ""typ"":""x', N'v';INSERT INTO authors (id, name, birthdate) VALUES (10001, 'Hidden Author', '2001-01-01');--"",
-                    ""val"":""x""
-                  }
-                ],
-                ""UserRoles"":[""authenticated""]
-            }";
-
-            string generatedToken = Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(header));
-
-            const string SESSION_CONFIG = $"session-context-config.{TestCategory.MSSQL}.json";
-            SetupSessionContextConfig(SESSION_CONFIG);
-
-            string[] args = [$"--ConfigFileName={SESSION_CONFIG}"];
-            using TestServer server = new(Program.CreateWebHostBuilder(args));
-            using HttpClient client = server.CreateClient();
-
-            HttpRequestMessage requestWithHeader = new(HttpMethod.Get, "api/Author");
-            requestWithHeader.Headers.Add(AuthenticationOptions.CLIENT_PRINCIPAL_HEADER, generatedToken);
-            requestWithHeader.Headers.Add(AuthorizationResolver.CLIENT_ROLE_HEADER, "authenticated");
-            HttpResponseMessage responseWithHeader = await client.SendAsync(requestWithHeader);
-            Assert.AreEqual(expected: HttpStatusCode.OK, actual: responseWithHeader.StatusCode);
-
-            HttpRequestMessage request = new(HttpMethod.Get, $"api/Author/id/10001");
-            HttpResponseMessage response = await client.SendAsync(request);
-            string responseBody = await response.Content.ReadAsStringAsync();
-
-            Assert.IsFalse(responseBody.Contains($"\"id\":10001"),
-                "The GET request should not return the invalid row.");
-            Assert.IsFalse(responseBody.Contains("Hidden Author"),
-                "The GET request should not return any information related to the invalid row.");
-        }
-
-        /// <summary>
         /// - Ensures an invalid EasyAuth header payload results in HTTP 401 Unauthorized response
         /// A correctly configured EasyAuth environment guarantees an EasyAuth payload for authenticated requests.
         /// - Ensures a missing EasyAuth header results in HTTP OK and User.IsAuthenticated == false.
@@ -474,40 +427,6 @@ namespace Azure.DataApiBuilder.Service.Tests.Authentication
 
                 context.Request.Scheme = "https";
             });
-        }
-
-        /// <summary>
-        /// Writes a MsSql runtime config that uses StaticWebApps EasyAuth and enables
-        /// set-session-context so requests exercise MsSqlQueryExecutor.GetSessionParamsQuery.
-        /// </summary>
-        private static void SetupSessionContextConfig(string configFileName)
-        {
-            TestHelper.SetupDatabaseEnvironment(TestCategory.MSSQL);
-            RuntimeConfigProvider configProvider =
-                TestHelper.GetRuntimeConfigProvider(TestHelper.GetRuntimeConfigLoader());
-            RuntimeConfig config = configProvider.GetConfig();
-
-            RuntimeConfig updatedConfig = config with
-            {
-                DataSource = config.DataSource! with
-                {
-                    Options = new Dictionary<string, object?>
-                    {
-                        // Matches MsSqlOptions.SetSessionContext (hyphenated naming policy).
-                        { "set-session-context", true }
-                    }
-                },
-                Runtime = config.Runtime! with
-                {
-                    Host = config.Runtime.Host! with
-                    {
-                        Authentication = new AuthenticationOptions(
-                            Provider: EasyAuthType.AppService.ToString(), Jwt: null)
-                    }
-                }
-            };
-
-            File.WriteAllText(configFileName, updatedConfig.ToJson());
         }
         #endregion
     }
